Character Block-In
This was a test go before i uploaded the version with the character file. (Because it merged the most recent with the one i was going to upload before my holiday). It seems to work perfectly aside from broken movement. (I was using wasd movement so its possible it didn't work due to no controller)