PlayerCharacter Rigged and animated!
The Player Character model has been fully rigged and animated! A Test animation is present within the player characters folder under the 'Animations' folder named 'TestAnimation'. I've taken the liberty to test the animation within the engine and it seems to work perfectly. Now all there is for it is to have modules set to apply a variety of animation imports, and triggers that will activate them.