Security camera and console
Added a new parent actor called BP_PowerDevice so end-of-circuit objects can inherit the same power system behaviors. Created a security camera that captures from the world and displays to a world-space widget on a console somewhere else in the map. Also made a flicker LED material that mimics worn down bands on an old screen.